pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gigaam"
version = "0.1.0"
description = "GigaAM: A package for audio modeling and ASR."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "GigaChat Team"}
]
dependencies = [
    "hydra-core==1.3.*",
    "numpy==2.*",
    "omegaconf==2.3.*",
    "onnx==1.19.*",
    "onnxruntime==1.23.*",
    "sentencepiece",
    "torch>=2.5,<2.9",
    "torchaudio>=2.5,<2.9",
    "tqdm",
]

[project.optional-dependencies]
longform = [
    "torch==2.8.*",
    "torchaudio==2.8.*",
    "pyannote.audio==4.0",
    "torchcodec==0.7",
    "numba>=0.62"
]
tests = [
    "pytest==9.*",
    "pytest-cov==7.*",
    "pytest-timeout==2.*",
    "scipy==1.*",
    "soundfile==0.13.*",
    "librosa==0.11.*"
]

[project.urls]
Repository = "https://github.com/salute-developers/GigaAM/"

[tool.setuptools]
include-package-data = true
license-files = ["LICENSE"]

[tool.setuptools.packages.find]
include = ["gigaam*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
timeout = 300
addopts = [
    "-v",
    "--cov=gigaam",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "--junitxml=junit.xml",
]

[tool.ruff]
line-length = 120
target-version = "py310"
exclude = [
    ".git",
    ".venv",
    "__pycache__",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.basedpyright]
pythonVersion = "3.10"
typeCheckingMode = "standard"
reportMissingImports = true
reportMissingTypeStubs = false
include = ["gigaam"]
exclude = [
    ".venv",
    "**/__pycache__",
    "tests",
]
